backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "dotenv>=0.9.9",
    "fastapi>=0.135.2",
    "PyJWT==2.12.1",
    "python-chess>=1.999",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.42.0",
]
